ultrasonic: reduce FRACTIONAL_US to 1
The timers that we're using to measure the echo signals are 16 bits
wide. The timeout is supposed to be set to 40ms, but with
FRACTIONAL_US == 4, this works out to 160,000 timer ticks, which the
timer isn't wide enough to count to. Instead, we were timing out at
160,000 modulo 65,536 timer ticks, or 28,928 ticks == 7232 us == 1.247
meters.

We could perhaps use some tricks to get more bits out of the timers,
like using an interrupt to count how many times the timer has rolled
over, but then if a timer rollover and a capture happen in close
proximity we will have to guess which happened first. I believe this
could be done, but it would be hard to test.

The best course of action here is to just reduce FRACTIONAL_US to 1 so
that the timeout value fits in 16 bits. This reduces our accuracy to
1us instead of 0.25us, but that's still about a sixth of a millimeter
which is good enough.